backend/requirements.txt
anthropic
openai
feedparser
requests
python-dotenv
colorama
